+++ telegram-bot/src/commonMain/kotlin/eu/vendeli/tgbot/types/internal/options/SavePreparedInlineMessageOptions.kt
@@ -1,5 +1,8 @@
 package eu.vendeli.tgbot.types.internal.options
 
+import kotlinx.serialization.Serializable
+
+@Serializable
 data class SavePreparedInlineMessageOptions(
     var allowUserChats: Boolean? = null,
     var allowBotChats: Boolean? = null,