+++ telegram-bot/src/commonMain/kotlin/eu/vendeli/tgbot/utils/serde/SerdeHelpers.kt
@@ -3,5 +3,5 @@ package eu.vendeli.tgbot.utils.serde
 import kotlinx.serialization.json.JsonElement
 import kotlinx.serialization.json.JsonPrimitive
 
-public val JsonElement.primitiveOrNull: JsonPrimitive?
+val JsonElement.primitiveOrNull: JsonPrimitive?
     get() = this as? JsonPrimitive